+++ template/target-libso.sh
@@ -1,0 +1,26 @@
+#!/bin/bash
+set -e
+
+libso_multi_version() {
+	local name_orig=$1
+	local name_v_v=$(echo ${name_orig} | grep -oE "lib[0-9a-zA-Z]+\.so\.[0-9]+\.[0-9]+")
+	local name_v=$(echo ${name_orig} | grep -oE "lib[0-9a-zA-Z]+\.so\.[0-9]+")
+	local name=$(echo ${name_orig} | grep -oE "lib[0-9a-zA-Z]+\.so")
+	if [[ ${name_orig} == ${name_v_v} ]]; then
+		ln -s ${name_orig} ${name_v}
+		ln -s ${name_v} ${name}
+	elif [[ ${name_orig} == ${name_v} ]]; then
+		ln -s ${name_orig} ${name}
+	fi
+}
+
+case $1 in
+multi-version)
+	shift
+	libso_multi_version ${1}
+	;;
+*)
+	echo >&2 "$0: ERROR: unknown $@"
+	exit 1
+	;;
+esac